rename: component 디렉토리 report 폴더구조 수정
