+++ src/components/report/dropdown/CurrentSolution.js
