+++ src/components/report/slide/CurrentSlide.js
