+++ src/components/report/slide/PastSlide.js
