+++ src/components/report/slide/SlideAnimation.js
@@ -13,17 +13,24 @@ import PastSlide from './PastSlide';
 export default function SlideAnimation() {
   return (
     <SwiperWrapper>
-      <Swiper slidesPerView= {2} // 화면에 보여질 slide 갯수
-              slidesOffsetAfter={40} // 오른쪽 끝에 여백 추가
+      <Swiper
+        slidesPerView={2} // 화면에 보여질 slide 갯수
+        slidesOffsetAfter={40} // 오른쪽 끝에 여백 추가
       >
         {/* TODO : 추후 props로 Slides컴포넌트 연결 예정 */}
-        <SwiperSlide><CurrentSlide /></SwiperSlide>
-        <SwiperSlide><PastSlide /></SwiperSlide>
-        <SwiperSlide><PastSlide /></SwiperSlide>
-        <SwiperSlide><PastSlide /></SwiperSlide>
+        <SwiperSlide>
+          <CurrentSlide />
+        </SwiperSlide>
+        <SwiperSlide>
+          <PastSlide />
+        </SwiperSlide>
+        <SwiperSlide>
+          <PastSlide />
+        </SwiperSlide>
+        <SwiperSlide>
+          <PastSlide />
+        </SwiperSlide>
       </Swiper>
     </SwiperWrapper>
   );
 }
-
-
